+++ Editor/Auditors/ShadersAuditor.cs
@@ -57,6 +57,36 @@ namespace Unity.ProjectAuditor.Editor.Auditors
         , IPreprocessBuildWithReport
 #endif
     {
+        static readonly IssueLayout k_ShaderLayout = new IssueLayout
+        {
+            category = IssueCategory.Shaders,
+            properties = new[]
+            {
+//                new PropertyDefinition { type = PropertyType.Severity},
+                new PropertyDefinition { type = PropertyType.Description, name = "Shader Name"},
+                new PropertyDefinition { type = PropertyType.Custom, format = PropertyFormat.Integer, name = "Actual Variants", longName = "Number of variants in the build" },
+                new PropertyDefinition { type = PropertyType.Custom + 1, format = PropertyFormat.Integer, name = "Passes", longName = "Number of Passes" },
+                new PropertyDefinition { type = PropertyType.Custom + 2, format = PropertyFormat.Integer, name = "Keywords", longName = "Number of Keywords" },
+                new PropertyDefinition { type = PropertyType.Custom + 3, format = PropertyFormat.Integer, name = "Render Queue" },
+                new PropertyDefinition { type = PropertyType.Custom + 4, format = PropertyFormat.Bool, name = "Instancing", longName = "GPU Instancing Support" },
+                new PropertyDefinition { type = PropertyType.Custom + 5, format = PropertyFormat.Bool, name = "SRP Batcher", longName = "SRP Batcher Compatible" }
+            }
+        };
+
+        static readonly IssueLayout k_ShaderVariantLayout = new IssueLayout
+        {
+            category = IssueCategory.ShaderVariants,
+            properties = new[]
+            {
+                new PropertyDefinition { type = PropertyType.Description, name = "Shader Name"},
+                new PropertyDefinition { type = PropertyType.Custom, format = PropertyFormat.Bool, name = "Compiled", longName = "Compiled at runtime by the player" },
+                new PropertyDefinition { type = PropertyType.Custom + 1, format = PropertyFormat.String, name = "Graphics API" },
+                new PropertyDefinition { type = PropertyType.Custom + 2, format = PropertyFormat.String, name = "Pass Name" },
+                new PropertyDefinition { type = PropertyType.Custom + 3, format = PropertyFormat.String, name = "Keywords" },
+                new PropertyDefinition { type = PropertyType.Custom + 4, format = PropertyFormat.String, name = "Requirements" }
+            }
+        };
+
         static readonly ProblemDescriptor k_ParseErrorDescriptor = new ProblemDescriptor
             (
             400000,
@@ -87,6 +117,12 @@ namespace Unity.ProjectAuditor.Editor.Auditors
         public IEnumerable<ProblemDescriptor> GetDescriptors()
         {
             yield return null;
+        }
+
+        public IEnumerable<IssueLayout> GetLayouts()
+        {
+            yield return k_ShaderLayout;
+            yield return k_ShaderVariantLayout;
         }
 
         public void Initialize(ProjectAuditorConfig config)