+++ Editor/IAuditor.cs
@@ -10,6 +10,8 @@ namespace Unity.ProjectAuditor.Editor
     {
         IEnumerable<ProblemDescriptor> GetDescriptors();
 
+        IEnumerable<IssueLayout> GetLayouts();
+
         void Initialize(ProjectAuditorConfig config);
 
         void Reload(string path);