+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 * Reduced UI managed allocations
 * Replaced tabs-like view selection with toolbar dropdown list
 * Removed *experimental* label from Allocation issues
+* Changed *Export* feature to be view-specific
 
 ## [0.4.2-preview] - 2021-02-01
 * Fixed detection of API calls using a derived type